requirements.txt
streamlit
langchain-groq
sentence-transformers
faiss-cpu
pandas
numpy
python-dotenv